Rename _apimeta to platform
